incr lookup extracted into gadget
