+++ crates/m3/src/gadgets/indexed_lookup/mod.rs
@@ -1,0 +1,1 @@
+pub mod incr;
